backend/server/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic<2.0.0
requests
supabase
python-dotenv
resend
